Consolidate TIR visitor traits into tir_visitor.rs
Move TirOptVisitor (formerly TirVisitor), walk functions, break-to
utilities, and visit_project_functions from optimize/visitor.rs into
the central tir_visitor.rs module.  Delete optimize/visitor.rs and
update all optimization passes to import from tir_visitor directly.

Also convert call_rewrite.rs manual traversals (sync_local_types_from_lets,
update_local_expr_types) to use TirRefVisitor / TirMutVisitor, removing
~130 lines of hand-written traversal boilerplate.
